stackproj/requirements.txt
Django==1.8.5
Pillow==2.3.0
PyYAML==3.11
amqp==1.4.7
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.21
celery==3.1.19
decorator==4.0.4
django-celery==3.1.17
django-rest-swagger==0.3.4
djangorestframework==3.3.0
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
kombu==3.0.29
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
ptyprocess==0.5
pytz==2015.7
simplegeneric==0.8.1
traitlets==4.0.0
wsgiref==0.1.2
